Moved Iterator to nested Class
While working on Windows, the previous approach had compilation errors on Linux
